Fix: wrapped된 시스템 콜을 사용하도록 수정

+++ mandatory/src/executor/set_standard_stream.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   set_standard_stream.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hyeyukim <hyeyukim@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: yeonhkim <yeonhkim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 16:24:28 by yeonhkim          #+#    #+#             */
-/*   Updated: 2023/01/25 16:06:34 by hyeyukim         ###   ########.fr       */
+/*   Updated: 2023/01/26 22:23:41 by yeonhkim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 #include "minishell.h"
 #include "parser.h"
 #include "env_manager.h"
+#include "wrapped_syscall.h"
 
 int	do_redirecting(t_redir *redir_list)
 {
 	int	fd;
-	int	err;
 	int	i;
 
 	i = 0;
@@ -29,7 +29,13 @@ int	do_redirecting(t_redir *redir_list)
 				|| redir_list[i].num == TOKEN_REDIR_IN_HERE)
 		{
 			fd = open(redir_list[i].str, O_RDONLY);
-			err = dup2(fd, STDIN_FILENO);
+			if (fd < 0)
+			{
+				handle_execute_error(ERR_EXE_NO_SUCH_FILE_OR_DIR, \
+												NULL, redir_list[i].str);
+				return (FAILURE);
+			}
+			w_dup2(fd, STDIN_FILENO);
 			if (redir_list[i].num == TOKEN_REDIR_IN_HERE)
 				unlink(redir_list[i].str);
 		}
@@ -37,22 +43,23 @@ int	do_redirecting(t_redir *redir_list)
 		{
 			fd = open(redir_list[i].str, \
 						O_WRONLY | O_CREAT | O_TRUNC, 0644);
-			err = dup2(fd, STDOUT_FILENO);
+			if (fd < 0)
+			{
+				handle_execute_error(ERR_EXE_NO_SUCH_FILE_OR_DIR, \
+												NULL, redir_list[i].str);
+				return (FAILURE);
+			}
+			w_dup2(fd, STDOUT_FILENO);
 		}
 		else if (redir_list[i].num == ERR_EXE_AMBIGUOUS_REDIR)
 		{
 			handle_execute_error(\
-			ERR_EXE_AMBIGUOUS_REDIR, NULL, redir_list[i].str);
-			return (ERR_EXE_AMBIGUOUS_REDIR);
-		}
-		if (err < 0)
-		{
-			exit_by_error("dup2 error");
-			return (ERR_EXE_SYSCALL_FAILED);
+				ERR_EXE_AMBIGUOUS_REDIR, NULL, redir_list[i].str);
+			return (FAILURE);
 		}
 		i++;
 	}
-	return (ERR_EXE_NONE);
+	return (SUCCESS);
 }
 
 static void	do_piping(int old_pipe_fd[2], int new_pipe_fd[2], \
@@ -64,30 +71,28 @@ static void	do_piping(int old_pipe_fd[2], int new_pipe_fd[2], \
 	if (nth == 1)
 	{
 		close(new_pipe_fd[P_READ]);
-		err[0] = dup2(new_pipe_fd[P_WRITE], STDOUT_FILENO);
+		err[0] = w_dup2(new_pipe_fd[P_WRITE], STDOUT_FILENO);
 		close(new_pipe_fd[P_WRITE]);
 	}
 	else if (nth == child_cnt)
 	{
-		err[0] = dup2(old_pipe_fd[P_READ], STDIN_FILENO);
+		err[0] = w_dup2(old_pipe_fd[P_READ], STDIN_FILENO);
 		close(old_pipe_fd[P_READ]);
 	}
 	else
 	{
 		close(new_pipe_fd[P_READ]);
-		err[0] = dup2(old_pipe_fd[P_READ], STDIN_FILENO);
-		err[1] = dup2(new_pipe_fd[P_WRITE], STDOUT_FILENO);
+		err[0] = w_dup2(old_pipe_fd[P_READ], STDIN_FILENO);
+		err[1] = w_dup2(new_pipe_fd[P_WRITE], STDOUT_FILENO);
 		close(old_pipe_fd[P_READ]);
 		close(new_pipe_fd[P_WRITE]);
 	}
-	if (err[0] < 0 || err[1] < 0)
-		exit_by_error("dup2 error");
 }
 
 void	set_standard_stream(t_pipeline *pl, t_redir *redir_list, int nth)
 {
 	if (pl && pl->pipe_exist)
 		do_piping(pl->old_pipe_fd, pl->new_pipe_fd, nth, pl->child_cnt);
-	if (do_redirecting(redir_list) > 0)
+	if (do_redirecting(redir_list) == FAILURE)
 		exit_program();
 }
